Rename `landing` to `onboarding`
