+++ lib/onboarding/existing_wallet/bloc/existing_wallet_state.dart
