+++ lib/onboarding/existing_wallet/view/confirm_passcode_screen.dart
@@ -1,19 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:genius_wallet/app/widgets/app_screen_with_header.dart';
-import 'package:genius_wallet/landing/existing_wallet/bloc/existing_wallet_bloc.dart';
+import 'package:genius_wallet/onboarding/existing_wallet/bloc/existing_wallet_bloc.dart';
+import 'package:genius_wallet/widgets/components/continue_button/isactive_false.g.dart';
 import 'package:genius_wallet/widgets/components/continue_button/isactive_true.g.dart';
 import 'package:genius_wallet/widgets/components/passcode_entry.g.dart';
 
-class CreatePasscodeScreen extends StatelessWidget {
-  const CreatePasscodeScreen({Key? key}) : super(key: key);
+class ConfirmPasscodeScreen extends StatelessWidget {
+  const ConfirmPasscodeScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: AppScreenWithHeader(
-        title: 'Set Passcode',
-        subtitle: 'Adds an extra layer of security when using the app',
+        title: 'Confirm your passcode',
+        subtitle: 'Enter your passcode once more',
         bodyWidgets: [
           ConstrainedBox(
             constraints: BoxConstraints(
@@ -26,7 +27,7 @@ class CreatePasscodeScreen extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.center,
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                const Text('Create Passcode'),
+                const Text('Confirm Passcode'),
                 SizedBox(
                   width: MediaQuery.of(context).size.width * 0.8,
                   height: 50,
@@ -50,9 +51,9 @@ class CreatePasscodeScreen extends StatelessWidget {
                 onPressed: () {
                   context
                       .read<ExistingWalletBloc>()
-                      .add(ChangeStep(step: FlowStep.confirmPasscode));
+                      .add(ChangeStep(step: FlowStep.importWallet));
                 },
-                child: IsactiveTrue(constraints),
+                child: IsactiveFalse(constraints),
               );
             },
           ),
