+++ lib/onboarding/existing_wallet/view/create_passcode_screen.dart
@@ -1,20 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:genius_wallet/app/widgets/app_screen_with_header.dart';
-import 'package:genius_wallet/landing/existing_wallet/bloc/existing_wallet_bloc.dart';
-import 'package:genius_wallet/widgets/components/continue_button/isactive_false.g.dart';
+import 'package:genius_wallet/onboarding/existing_wallet/bloc/existing_wallet_bloc.dart';
 import 'package:genius_wallet/widgets/components/continue_button/isactive_true.g.dart';
 import 'package:genius_wallet/widgets/components/passcode_entry.g.dart';
 
-class ConfirmPasscodeScreen extends StatelessWidget {
-  const ConfirmPasscodeScreen({super.key});
+class CreatePasscodeScreen extends StatelessWidget {
+  const CreatePasscodeScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: AppScreenWithHeader(
-        title: 'Confirm your passcode',
-        subtitle: 'Enter your passcode once more',
+        title: 'Set Passcode',
+        subtitle: 'Adds an extra layer of security when using the app',
         bodyWidgets: [
           ConstrainedBox(
             constraints: BoxConstraints(
@@ -27,7 +26,7 @@ class ConfirmPasscodeScreen extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.center,
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                const Text('Confirm Passcode'),
+                const Text('Create Passcode'),
                 SizedBox(
                   width: MediaQuery.of(context).size.width * 0.8,
                   height: 50,
@@ -51,9 +50,9 @@ class ConfirmPasscodeScreen extends StatelessWidget {
                 onPressed: () {
                   context
                       .read<ExistingWalletBloc>()
-                      .add(ChangeStep(step: FlowStep.importWallet));
+                      .add(ChangeStep(step: FlowStep.confirmPasscode));
                 },
-                child: IsactiveFalse(constraints),
+                child: IsactiveTrue(constraints),
               );
             },
           ),
