+++ lib/onboarding/new_wallet/bloc/new_wallet_event.dart
