+++ lib/onboarding/new_wallet/routes/new_wallet_flow.dart
@@ -1,10 +1,10 @@
 import 'package:flow_builder/flow_builder.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:genius_wallet/landing/new_wallet/bloc/new_wallet_bloc.dart';
-import 'package:genius_wallet/landing/new_wallet/view/backup_phrase_screen.dart';
-import 'package:genius_wallet/landing/new_wallet/view/recovery_phrase_screen.dart';
-import 'package:genius_wallet/landing/new_wallet/view/verify_recovery_phrase_screen.dart';
+import 'package:genius_wallet/onboarding/new_wallet/bloc/new_wallet_bloc.dart';
+import 'package:genius_wallet/onboarding/new_wallet/view/backup_phrase_screen.dart';
+import 'package:genius_wallet/onboarding/new_wallet/view/recovery_phrase_screen.dart';
+import 'package:genius_wallet/onboarding/new_wallet/view/verify_recovery_phrase_screen.dart';
 
 class NewWalletFlow extends StatelessWidget {
   const NewWalletFlow({Key? key}) : super(key: key);
