+++ lib/onboarding/view/keystore_tab_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:genius_wallet/landing/widgets/paste_field.dart';
+import 'package:genius_wallet/onboarding/widgets/paste_field.dart';
 import 'package:genius_wallet/widgets/components/text_entry_field_widget.g.dart';
 import 'package:genius_wallet/widgets/text_form_field_logic.g.dart';
