+++ lib/onboarding/widgets/recovery_words.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:genius_wallet/landing/new_wallet/bloc/new_wallet_bloc.dart';
+import 'package:genius_wallet/onboarding/new_wallet/bloc/new_wallet_bloc.dart';
 import 'package:genius_wallet/widgets/components/recoveryword.g.dart';
 
 class RecoveryWords extends StatelessWidget {
