+++ lib/widgets/components/custom/type_create_custom.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:genius_api/genius_api.dart';
-import 'package:genius_wallet/landing/new_wallet/bloc/new_wallet_bloc.dart';
-import 'package:genius_wallet/landing/new_wallet/routes/new_wallet_flow.dart';
+import 'package:genius_wallet/onboarding/new_wallet/bloc/new_wallet_bloc.dart';
+import 'package:genius_wallet/onboarding/new_wallet/routes/new_wallet_flow.dart';
 
 class TypeCreateCustom extends StatefulWidget {
   final Widget? child;
