+++ lib/widgets/components/custom/type_existing_custom.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:genius_wallet/landing/existing_wallet/bloc/existing_wallet_bloc.dart';
-import 'package:genius_wallet/landing/existing_wallet/routes/existing_wallet_flow.dart';
+import 'package:genius_wallet/onboarding/existing_wallet/bloc/existing_wallet_bloc.dart';
+import 'package:genius_wallet/onboarding/existing_wallet/routes/existing_wallet_flow.dart';
 
 class TypeExistingCustom extends StatefulWidget {
   final Widget? child;
